+++ packages/snap-preact-demo/webpack.dev.js
@@ -1,14 +1,26 @@
-const PreactRefreshPlugin = require('@prefresh/webpack');
 const { merge } = require('webpack-merge');
 const common = require('./webpack.common.js');
 const path = require('path');
 
-module.exports = merge(common, {
+const universal = merge(common, {
 	mode: 'development',
-	target: 'web',
-	devtool: 'source-map',
+	entry: './src/universal.js',
+	output: {
+		filename: 'universal.bundle.js',
+		chunkFilename: 'snap.universal.chunk.[fullhash:8].[id].js',
+	},
+	target: 'browserslist',
+	module: {
+		rules: [
+			{
+				test: /\.(js|jsx)$/,
+				exclude: /node_modules/,
+				use: ['babel-loader'],
+			},
+		],
+	},
 	devServer: {
-		https: false,
+		https: true,
 		port: 4444,
 		hot: true,
 		allowedHosts: 'all',
@@ -24,5 +36,27 @@ module.exports = merge(common, {
 			publicPath: '/dist/',
 		},
 	},
-	plugins: [new PreactRefreshPlugin()],
 });
+
+const modern = merge(common, {
+	mode: 'development',
+	entry: './src/index.js',
+	output: {
+		filename: 'bundle.js',
+		chunkFilename: 'snap.chunk.[fullhash:8].[id].js',
+		publicPath: '/dist/',
+	},
+	module: {
+		rules: [
+			{
+				test: /\.(js|jsx)$/,
+				exclude: /node_modules/,
+				use: ['babel-loader'],
+			},
+		],
+	},
+	target: 'web',
+	devtool: 'source-map',
+});
+
+module.exports = [universal, modern];